requirements.txt
# Python 3.5
Django==1.11.7
django-simple-captcha
